requirements.txt
python-scrapyd-api
flask
pymongo
aiohttp
scrapyd
pywin32
scrapy
